+++ src/mcp_tools.py
@@ -38,8 +38,6 @@ from evo_mcp.tools import (
     register_object_builder_tools,
     register_file_tools,
     register_instance_users_admin_tools,
-    register_duplicate_tools,
-    register_object_compare_tools,
 )
 
 # Get transport mode from environment variable
@@ -96,11 +94,10 @@ register_general_tools(mcp)
 
 if TOOL_FILTER in ["all", "admin"]:
     # Admin Agent: Workspace and instance management tools
-    # Includes: workspace creation, snapshots, duplication, permissions management
+    # Includes: workspace creation, snapshots, duplication, permissions management,
+    # and cross-workspace comparison analysis
     register_admin_tools(mcp)
     register_instance_users_admin_tools(mcp)
-    register_duplicate_tools(mcp)
-    register_object_compare_tools(mcp)
 if TOOL_FILTER in ["all", "data"]: #  "data_agent"
     # register_data_tools(mcp)
     register_filesystem_tools(mcp)
